Add app skeleton, show creation of micro Spring context in test
The `sql-backed-persistence` module has a test that needs to stand up a
Spring context (in order to get access to the implementation of the JPA
repository that Spring provides).

Rather than relying on the overall Spring context (which is defined in a
separate module anyway, as `sql-backed-persistence` is not intended to
define a stand-alone application), a tiny Spring context is created
where it's needed, in the SqlBackedEventRepositoryTest.